1.6.2.12 - Loading.
